+++ android/app/src/main/java/org/fossasia/openevent/events/SessionDownloadEvent.java
@@ -1,0 +1,22 @@
+package org.fossasia.openevent.events;
+
+/**
+ * User: MananWason
+ * Date: 8/4/2015
+ */
+public class SessionDownloadEvent {
+    private boolean state;
+
+    public SessionDownloadEvent(boolean state) {
+
+        this.state = state;
+    }
+
+    public boolean isState() {
+        return state;
+    }
+
+    public void setState(boolean state) {
+        this.state = state;
+    }
+}